+++ src/PersonJsonFixtures.php
@@ -1,0 +1,28 @@
+<?php
+
+/*
+ * This file is part of the xAPI package.
+ *
+ * (c) Christian Flothmann <christian.flothmann@xabbuh.de>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace XApi\Fixtures\Json;
+
+/**
+ * JSON encoded xAPI person fixtures.
+ *
+ * These fixtures are borrowed from the
+ * {@link https://github.com/adlnet/xAPI_LRS_Test Experience API Learning Record Store Conformance Test} package.
+ */
+class PersonJsonFixtures extends JsonFixtures
+{
+    public const DIRECTORY = 'Person';
+
+    public static function getTypicalPerson(): string
+    {
+        return self::load('typical_person');
+    }
+}