+++ src/StatementJsonFixtures.php
@@ -22,8 +22,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a minimal valid statement.
-     *
-     * @return string
      */
     public static function getMinimalStatement(): string
     {
@@ -47,8 +45,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a statement with a group as an actor.
-     *
-     * @return string
      */
     public static function getStatementWithGroupActor(): string
     {
@@ -57,8 +53,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a statement with a group that has no members as an actor.
-     *
-     * @return string
      */
     public static function getStatementWithGroupActorWithoutMembers(): string
     {
@@ -67,8 +61,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a statement including a reference to another statement.
-     *
-     * @return string
      */
     public static function getStatementWithStatementRef(): string
     {
@@ -77,8 +69,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a statement including a result.
-     *
-     * @return string
      */
     public static function getStatementWithResult(): string
     {
@@ -87,8 +77,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a statement including a sub statement.
-     *
-     * @return string
      */
     public static function getStatementWithSubStatement(): string
     {
@@ -97,8 +85,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a statement including an agent authority.
-     *
-     * @return string
      */
     public static function getStatementWithAgentAuthority(): string
     {
@@ -107,8 +93,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a statement including a group authority.
-     *
-     * @return string
      */
     public static function getStatementWithGroupAuthority(): string
     {
@@ -122,8 +106,6 @@ class StatementJsonFixtures extends JsonFixtures
 
     /**
      * Loads a collection of statements.
-     *
-     * @return string
      */
     public static function getStatementCollection(): string
     {